Add test for service chain next-hop manipulation
Added a test to verify that the service chain routes are recomputed after the
service chain next-hop goes from ECMP to non-ECMP. This is to emulate cases
where the service instance is configured with two interfaces running VRRP with
a VIP assigned. The VIP being the service chain next-hop.
The VIP will be added for active paths with Local preference 200
and standby paths with local preference 100. It is possible to have both stanby
with a local preference 100, it should however recover when one becomes active.

Change-Id: I2934ddc0c94561125bd5eb05390bc5ff36239414