Fix label allocation for multicast
During BFD scale tests a crash was observed when attempting to
free an MPLS label, It turns out the label for a local interface
was allocated a label in the multicast range.
The root cause - during the allocation of label blocks per
control node the last label in the block wasn't actually
reserved. This label was allocated to locally on the agent for a VMI.
When attempting to free the label the check for the multicast range
suceeds, at this point it attempts to look up the vrf name from the
label data (changes made in bug#1724114), the label data is not set
in this case and causes the crash.

Fix: Correct the label block allocation for multicast to reserve all
labels in the range.

Closes-Bug: 1798483